Initial commit
The OpenShift plugin for Jenkins is an aggregator of the plugins that
Red Hat supports in its distribution of Jenkins. It includes the
OpenShift login, client, and sync plugins as direct dependencies as
well as the upstream Kubernetes plugin for Jenkins. Due to the current
state of the OpenShift dependencies, this plugin is based on a Jenkins
version that has currently reached end of life.

The plugin is only meant to be used within a container running on
OpenShift. The container image is built on Red Hat's Konflux instance
using a newer, supported version of Jenkins. The additional plugins in
`.konflux/plugins.txt` are installed by the plugin manager tool. In the
future, these plugins should be added as direct dependencies in this
plugin's `pom.xml`.