Support Multiple `oc` Versions
Update Jenkins image to support multiple `oc` client versions. The
`values-client.yaml` configures Jenkins to switch the version of oc
used within a pipeline (via UI or the `tools` DSL in a Jenkinsfile). A
default version of `oc` is set by extending the `PATH` environment
variable on the Jenkins built in node using the special `ENVVAR+EXTRA`
syntax that only Jenkins supports.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>